pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tax-resolution-medallion"
version = "1.0.0"
description = "Medallion architecture for tax resolution data processing"
authors = [
    {name = "Tax Resolution Team"}
]
requires-python = ">=3.9"
dependencies = [
    "dagster>=1.5.0",
    "dagster-webserver>=1.5.0",
    "dagster-cloud",
    "supabase>=2.3.0",
    "httpx>=0.26.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["dagster_pipeline*"]
exclude = ["backend*", "TRA_API*", "supabase*", "ExistingDocs*"]

[tool.dagster]
module_name = "dagster_pipeline"

